backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sentence-transformers
python-docx
pypdf2
python-multipart
numpy
scikit-learn
